Refactor code
- Use same table as before for amm_events
- Decouple more code
- Update tests
- Move interface
