+++ rotkehlchen/chain/ethereum/interfaces/ammswap/graph.py
@@ -185,3 +185,41 @@ TOKEN_DAY_DATAS_QUERY = (
     }}}}
     """
 )
+
+LIQUIDITY_POSITIONS_QUERY = (
+    """
+    liquidityPositions
+    (
+        first: $limit,
+        skip: $offset,
+        where: {{
+            user_in: $addresses,
+            liquidityTokenBalance_gt: $balance,
+        }}
+    ) {{
+        id
+        liquidityTokenBalance
+        pair {{
+            id
+            reserve0
+            reserve1
+            token0 {{
+                id
+                decimals
+                name
+                symbol
+            }}
+            token1 {{
+                id
+                decimals
+                name
+                symbol
+            }}
+            totalSupply
+        }}
+        user {{
+            id
+        }}
+    }}}}
+    """
+)
