+++ rotkehlchen/chain/ethereum/interfaces/ammswap/utils.py
@@ -18,8 +18,8 @@ log = logging.getLogger(__name__)
 
 
 SUBGRAPH_REMOTE_ERROR_MSG = (
-    "Failed to request the Uniswap subgraph due to {error_msg}. "
-    "All Uniswap balances and historical queries are not functioning until this is fixed. "  # noqa: E501
+    "Failed to request the {location} subgraph due to {error_msg}. "
+    "All {location} balances and historical queries are not functioning until this is fixed. "  # noqa: E501
     "Probably will get fixed with time. If not report it to rotki's support channel"  # noqa: E501
 )
