+++ rotkehlchen/chain/ethereum/modules/uniswap/graph.py
@@ -1,43 +1,3 @@
-# Get balances queries
-
-LIQUIDITY_POSITIONS_QUERY = (
-    """
-    liquidityPositions
-    (
-        first: $limit,
-        skip: $offset,
-        where: {{
-            user_in: $addresses,
-            liquidityTokenBalance_gt: $balance,
-        }}
-    ) {{
-        id
-        liquidityTokenBalance
-        pair {{
-            id
-            reserve0
-            reserve1
-            token0 {{
-                id
-                decimals
-                name
-                symbol
-            }}
-            token1 {{
-                id
-                decimals
-                name
-                symbol
-            }}
-            totalSupply
-        }}
-        user {{
-            id
-        }}
-    }}}}
-    """
-)
-
 # Get trades queries
