+++ rotkehlchen/tests/db/test_db.py
@@ -113,7 +113,7 @@ TABLES_AT_INIT = [
     'xpubs',
     'xpub_mappings',
     'amm_swaps',
-    'amm_events',
+    'uniswap_events',
     'eth2_deposits',
     'eth2_daily_staking_details',
     'adex_events',
