+++ rotkehlchen/tests/unit/uniswap/test_get_unknown_asset_price.py
@@ -39,6 +39,7 @@ def test_pagination(
         graph_query_limit,
         no_requests,
         mock_graph_query_limit,  # pylint: disable=unused-argument
+        mock_amm_graph_query_limit,  # pylint: disable=unused-argument
 ):
     """Test an extra graph request is done when the number of items in the
     response equals GRAPH_QUERY_LIMIT.
